+++ OUDS/Core/Components/Sources/Controls/Radio/Internal/RadioIndicatorModifiers.swift
@@ -58,7 +58,7 @@ private struct RadioIndicatorForegroundModifier: ViewModifier {
 
     func body(content: Content) -> some View {
         content
-            .oudsForegroundColor(appliedColor)
+            .foregroundColor(appliedColor)
     }
 
     // MARK: - Colors
@@ -185,7 +185,7 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
 
     func body(content: Content) -> some View {
         content
-            .oudsBorder(style: theme.borders.styleDefault,
+            .border(style: theme.borders.styleDefault,
                         width: appliedBorderWidth,
                         radius: appliedBorderRadius,
                         color: appliedColor)