+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/TextInputBackgroundModifier.swift
@@ -29,7 +29,7 @@ struct TextInputBackgroundModifier: ViewModifier {
 
     func body(content: Content) -> some View {
         if let color {
-            content.oudsBackground(color)
+            content.background(color)
         } else {
             content
         }