+++ OUDS/Core/Components/Sources/Controls/TextInput/Internal/TextInputBorderModifier.swift
@@ -34,7 +34,7 @@ struct TextInputBorderModifier: ViewModifier {
                 content
             } else {
                 content
-                    .oudsBorder(style: theme.borders.styleDefault,
+                    .border(style: theme.borders.styleDefault,
                                 width: size,
                                 radius: cornerRadius,
                                 color: outlinedColor)
@@ -42,7 +42,7 @@ struct TextInputBorderModifier: ViewModifier {
         } else {
             if status == .readOnly {
                 content
-                    .oudsBorder(style: theme.borders.styleDefault,
+                    .border(style: theme.borders.styleDefault,
                                 width: theme.textInput.borderWidthDefault,
                                 radius: cornerRadius,
                                 color: theme.colors.borderMuted)