+++ OUDS/Core/Components/Sources/Indicators/Tag/Internal/Tag/TagLabelStyle+TagLabel.swift
@@ -38,7 +38,7 @@ struct TagLabel: View {
                     .labelStrongSmall(theme)
             }
         }
-        .oudsForegroundColor(color)
+        .foregroundColor(color)
     }
 
     // MARK: Helpers