+++ OUDS/Core/Components/Sources/Layouts/Divider/OUDSVerticalDivider.swift
@@ -74,11 +74,31 @@ public struct OUDSVerticalDivider: View {
     // MARK: - Body
 
     public var body: some View {
-        Divider().oudsVerticalDivider(color: color)
+        Divider().verticalDivider(color: color)
     }
 }
 
 extension Divider {
+    /// Set the color to the vertical divider and adjust the thickness (width)
+    /// automatically according to token value.
+    ///
+    /// ```swift
+    /// HStack {
+    ///     Text("Hello wolrd!")
+    ///        .verticalDivider(color: .brandPrimary)
+    ///
+    ///     Text("Happy to see you")
+    /// }
+    /// ```
+    ///
+    /// - Parameter color: the color of the divider, `OUDSDividerColor.default` by default
+    ///
+    /// **Remark: Vertical means vertical line (content separator in a HStack)**
+    @MainActor
+    public func verticalDivider(color: OUDSDividerColor) -> some View {
+        modifier(DividerModifier(orientation: .vertical, dividerColor: color))
+    }
+
     /// Set the color to the vertical divider and adjust the thickness (width)
     /// automatically according to token value.
     ///
@@ -94,8 +114,9 @@ extension Divider {
     /// - Parameter color: the color of the divider, `OUDSDividerColor.default` by default
     ///
     /// **Remark: Vertical means vertical line (content separator in a HStack)**
+    @available(*, deprecated, renamed: "verticalDivider(color:)", message: "Use verticalDivider(color:) instead")
     @MainActor
     public func oudsVerticalDivider(color: OUDSDividerColor) -> some View {
-        modifier(DividerModifier(orientation: .vertical, dividerColor: color))
+        verticalDivider(color: color)
     }
 }